+++ mitra-config/src/federation.rs
@@ -1,7 +1,21 @@
 use serde::Deserialize;
 
-#[derive(Clone, Default, Deserialize)]
+fn default_federation_enabled() -> bool { true }
+
+#[derive(Clone, Deserialize)]
 pub struct FederationConfig {
-    pub proxy_url: Option<String>,
-    pub onion_proxy_url: Option<String>,
+    #[serde(default = "default_federation_enabled")]
+    pub enabled: bool,
+    pub(super) proxy_url: Option<String>,
+    pub(super) onion_proxy_url: Option<String>,
+}
+
+impl Default for FederationConfig {
+    fn default() -> Self {
+        Self {
+            enabled: default_federation_enabled(),
+            proxy_url: None,
+            onion_proxy_url: None,
+        }
+    }
 }